pyproject.toml
[project]
name = "AWS"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12,<3.13"
dependencies = [
    "django>=5.2.7",
    "django-filter>=25.2",
    "django-restframework>=0.0.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt[blacklist]>=5.5.1",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.28.0",
    "drf-yasg>=1.21.11",
    "gunicorn>=23.0.0",
    "whitenoise",
    "psycopg[binary]",
    "ruff>=0.14.0",
    "itsdangerous>=2.2.0",
    "black>=25.9.0",
    "isort>=7.0.0",
    "mypy>=1.18.2",
    "python-json-logger>=4.0.0",
    "google-genai>=1.46.0",
]

[tool.black]
line-length = 88
target-version = ['py310']
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip = ["migrations", "__pycache__", ".venv"]

[dependency-groups]
dev = [
    "django-stubs>=5.2.7",
]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true
strict_optional = true

[tool.django-stubs]
django_settings_module = "AWS.settings"
strict_settings = false

